requirements.txt
datasets==3.4.1
einops==0.8.1
einx==0.3.0
Flask==3.1.0
flask_cors==5.0.1
gradio==5.22.0
librosa==0.11.0
numpy
omegaconf==2.3.0
packaging==24.2
pydub==0.25.1
Requests==2.32.3
safetensors==0.5.3
scipy==1.15.2
semantic_text_splitter==0.25.0
soundfile==0.12.1
soxr==0.5.0.post1
torch>=2.5.1
torchaudio>=2.5.1
tqdm==4.66.5
transformers==4.46.2
tritonclient==2.55.0
demoji
chatterbox-tts
